+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # domino x.x.x (not yet released)
 * Implement WindowTimers interface on Window. (#72)
 * Factor out the NavigatorID interface and make more spec-compliant.
+* Implement `HTMLTemplateElement` and parse `<template>` tags.
 
 # domino 1.0.23 (30 Jan 2016)
 * Fix `CSSStyleDeclaration#setProperty`. (#71)